UseContext for event module
