+++ src/components/EventModule/index.js
@@ -1,0 +1,1 @@
+export { default } from './EventModuleWithContext';
